Fix regex for nick highlighting
As Markb pointed out in #quassel the nick highlighting code was flawed. It used ([\W\D]|^) which would match any non-word char, any non-digit char and the beginning of a new line, same at the other and of the nick. Using the [\b] char class fixes that, as it matches wordboundaries.